Set correct TERM
With the wrong TERM, only limited colors are used. Change is obvious in
circumstances such as syntax highlighting in vim or PS1 of bash.
On the other hand, however, probably from this change, a glitch happens.
When editing text in vim, text or message on the status bar blinks
sometimes. I cannot reproduce the bug at the time, and also I'm not sure
that the bug is due to the change of TERM.